requirements.txt
python-dotenv==1.0.0
azure-identity==1.14.1
openai==1.3.3
httpx<0.28.0
fastapi==0.110.2
uvicorn[standard]==0.29.0
requests==2.31.0
streamlit==1.33.0
pydantic>=2.6.0
pytest==8.3.2